+++ common/src/main/java/pepjebs/mapatlases/client/screen/DimensionBookmarkButton.java
@@ -1,6 +1,7 @@
 package pepjebs.mapatlases.client.screen;
 
 import com.mojang.blaze3d.vertex.PoseStack;
+import net.mehvahdjukaar.candlelight.api.VirtualOverride;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.Tooltip;
@@ -14,15 +15,13 @@ import pepjebs.mapatlases.MapAtlasesMod;
 import pepjebs.mapatlases.client.MapAtlasesClient;
 import pepjebs.mapatlases.config.MapAtlasesClientConfig;
 
-import static pepjebs.mapatlases.client.MapAtlasesClient.ATLAS_BACKGROUND_TEXTURE;
-
-public class DimensionBookmarkButton extends BookmarkButton {
+public class DimensionBookmarkButton extends AtlasButton {
 
     private static final int BUTTON_H = 18;
     private static final int BUTTON_W = 24;
 
     private final ResourceKey<Level> dimension;
-    private final ResourceLocation sprite;
+    private final ResourceLocation dimensionIconSprite;
 
 
     protected DimensionBookmarkButton(int pX, int pY, ResourceKey<Level> dimension, AtlasOverviewScreen screen) {
@@ -31,17 +30,16 @@ public class DimensionBookmarkButton extends BookmarkButton {
         this.dimension = dimension;
         this.setTooltip(createTooltip());
         ResourceLocation res = MapAtlasesMod.res("dimensions/" + dimension.location().getPath());
-        //check for missing
         if (Minecraft.getInstance().getGuiSprites().getSprite(res) ==
                 Minecraft.getInstance().getGuiSprites().getSprite(MapAtlasesMod.res("missing"))) {
             res = MapAtlasesMod.res("dimension/overworld");
         }
-        this.sprite = res;
+        this.dimensionIconSprite = res;
     }
 
     @Override
     public Tooltip createTooltip() {
-        return Tooltip.create(Component.literal(AtlasOverviewScreen.getReadableName(dimension.location())));
+        return Tooltip.create(Component.literal(AtlasScreenUtils.getReadableName(dimension.location())));
     }
 
     public ResourceKey<Level> getDimension() {
@@ -57,8 +55,8 @@ public class DimensionBookmarkButton extends BookmarkButton {
             pose.translate(0, 0, 2);
         }
         super.renderWidget(pGuiGraphics, pMouseX, pMouseY, pPartialTick);
-        pGuiGraphics.blitSprite(sprite,
-                this.getX() + 4, this.getY() , 16, 16);
+        pGuiGraphics.blitSprite(dimensionIconSprite,
+                this.getX() + 4, this.getY(), 16, 16);
         pose.popPose();
 
     }
@@ -69,7 +67,7 @@ public class DimensionBookmarkButton extends BookmarkButton {
         parentScreen.selectDimension(dimension);
     }
 
-    //@Override
+    @VirtualOverride("neoforge")
     public void onClick(double mouseX, double mouseY, int button) {
         onClick(mouseX, mouseY);
     }