+++ common/src/main/java/pepjebs/mapatlases/client/screen/SliceArrowButton.java
@@ -1,6 +1,7 @@
 package pepjebs.mapatlases.client.screen;
 
 import com.mojang.blaze3d.vertex.PoseStack;
+import net.mehvahdjukaar.candlelight.api.VirtualOverride;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.resources.sounds.SimpleSoundInstance;
 import net.minecraft.client.sounds.SoundManager;
@@ -11,7 +12,7 @@ import pepjebs.mapatlases.config.MapAtlasesClientConfig;
 
 import java.util.TreeSet;
 
-public class SliceArrowButton extends BookmarkButton {
+public class SliceArrowButton extends AtlasButton {
 
     private static final int BUTTON_H = 5;
     private static final int BUTTON_W = 8;
@@ -84,7 +85,7 @@ public class SliceArrowButton extends BookmarkButton {
         else this.parentScreen.increaseSlice();
     }
 
-    //@Override
+    @VirtualOverride("neoforge")
     public void onClick(double mouseX, double mouseY, int button) {
         onClick(mouseX, mouseY);
     }