+++ common/src/main/java/pepjebs/mapatlases/client/screen/SliceBookmarkButton.java
@@ -2,6 +2,7 @@ package pepjebs.mapatlases.client.screen;
 
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
+import net.mehvahdjukaar.candlelight.api.VirtualOverride;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.Tooltip;
 import net.minecraft.network.chat.Component;
@@ -10,7 +11,7 @@ import pepjebs.mapatlases.client.MapAtlasesClient;
 import pepjebs.mapatlases.config.MapAtlasesClientConfig;
 import pepjebs.mapatlases.utils.Slice;
 
-public class SliceBookmarkButton extends BookmarkButton {
+public class SliceBookmarkButton extends AtlasButton {
 
     private static final int BUTTON_H = 21;
     private static final int BUTTON_W = 27;
@@ -24,7 +25,7 @@ public class SliceBookmarkButton extends BookmarkButton {
     protected SliceBookmarkButton(int pX, int pY, Slice slice, AtlasOverviewScreen screen) {
         super(pX, pY, BUTTON_W, BUTTON_H, screen, MapAtlasesClient.SLICE_BOOKMARK_SPRITE, MapAtlasesClient.SLICE_BOOKMARK_SPRITE);
         this.slice = slice;
-        this.selected = false;
+        this.setSelected(false);
         this.setTooltip(createTooltip());
     }
 
@@ -87,7 +88,7 @@ public class SliceBookmarkButton extends BookmarkButton {
 
     }
 
-    //@Override
+    @VirtualOverride("neoforge")
     public void onClick(double mouseX, double mouseY, int button) {
         onClick(mouseX, mouseY);
     }