+++ common/src/main/java/pepjebs/mapatlases/integration/moonlight/MoonlightCompat.java
@@ -22,13 +22,14 @@ import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.HitResult;
 import org.jetbrains.annotations.Nullable;
 import pepjebs.mapatlases.MapAtlasesMod;
+import pepjebs.mapatlases.utils.CustomDecorationHolder;
 import pepjebs.mapatlases.utils.DecorationHolder;
 import pepjebs.mapatlases.utils.MapDataHolder;
 
 import java.util.*;
 
 public class MoonlightCompat {
-    private static final TagKey<MLMapDecorationType<?, ?>> NOT_ON_ATLAS = TagKey.create(MapDataRegistry.REGISTRY_KEY,
+    private static final TagKey<MLMapDecorationType<?, ?>> NOT_ON_ATLAS = TagKey.create(MapDataRegistry.MAP_DECORATION_REGISTRY_KEY,
             MapAtlasesMod.res("no_button_on_atlas"));
 
     private static final ResourceLocation PIN_TYPE_ID = MapAtlasesMod.res("pin");
@@ -46,10 +47,10 @@ public class MoonlightCompat {
         }
     }
 
-    public static Collection<DecorationHolder> getCustomDecorations(MapDataHolder map) {
+    public static Collection<CustomDecorationHolder> getCustomDecorations(MapDataHolder map) {
         return ((ExpandedMapData) map.data).ml$getCustomDecorations().entrySet().stream()
                 .filter(e -> !e.getValue().getType().is(NOT_ON_ATLAS))
-                .map(a -> new DecorationHolder(a.getValue(), a.getKey(), map)).toList();
+                .map(a -> DecorationHolder.custom(a.getValue(), a.getKey(), map)).toList();
     }
 
     public static void addDecoration(Level level, MapItemSavedData data, BlockPos pos, ResourceLocation id, @Nullable Component name) {